Remove unused message column from createTable statement, #671
* still reading from the old message column if it is defined and has data
* using same column definition caching as for some other things
